+++ src/pages/Home/Gallery/GalleryHOC.tsx
@@ -1,21 +1,8 @@
-import React from "react";
 import { useRecoilValue } from "recoil";
 import Gallery from ".";
 import useGetThumbnails from "../../../hooks/Gallery/useGetThumbnails";
 import { galleryArts } from "../../../store/gallery";
-import { TArtImageResponse, TThumbnail } from "./types";
-
-export type TImagesRes = {
-  data?: TArtImageResponse;
-  isLoading: boolean;
-  setNextQuery: React.Dispatch<
-    React.SetStateAction<{
-      nextRevisionDateIdx?: string;
-      nextIdx: number;
-    } | null>
-  >;
-  thumbnailsAll?: TThumbnail[];
-};
+import { TImagesRes } from "./types";
 
 const GalleryHOC = () => {
   const thumbnailsAll = useRecoilValue(galleryArts);
